# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Nota API
  version: 0.1.0
  description: >-
    HTTP API for the Nota marketplace: the public carnet of notarial-act bids for
    Quebec City. Every offer is revalidated server-side through @nota/domain; the
    client's tier, premium and total are never trusted, and a bid marked
    `anonyme` never leaks its `nom`.


    NOTE: this specification is kept in sync with `apps/api/src/handler.js` **by
    hand** for now. When a route or field changes there, update it here in the
    same change (see AGENTS.md).
  license:
    name: UNLICENSED
servers:
  - url: /api
    description: Same-origin path fronted by CloudFront (production).
  - url: http://localhost:8788
    description: Local dev server (apps/api/local-server.js).
# Public intake and sign-in requests accept the standard Accept-Language header.
# Supported languages: fr and en, honoring quality weights; unsupported -> fr.
# New recipients retain this language for notifications. Anonymous requests never
# overwrite a saved preference; verified sign-in and preference updates can.
paths:
...
  /signing-beta/sessions:
    get:
      summary: Poll the current retained-bid rehearsal using a recently email-verified bearer token
      operationId: signingBetaRead
      security: [{ notarySession: [] }, { clientToken: [] }]
      parameters:
        - { name: bidId, in: query, required: true, schema: { type: string } }
        - { name: dateISO, in: query, required: true, schema: { type: string, format: date } }
      responses:
        '200':
          description: Current session, own role, peer-only SDP, own one-use proof challenge and admitted TURN credentials.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SigningBetaResponse' }
        default: { $ref: '#/components/responses/SigningBetaError' }
    post:
      summary: Notary creates a thirty-minute test-document rehearsal for a retained bid
      operationId: signingBetaCreate
      security: [{ notarySession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - { $ref: '#/components/schemas/SigningBetaCommand' }
                - { required: [publicKeyJwk] }
      responses:
        '201':
          description: Immutable test document and empty two-person room created; existing active room conflicts.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SigningBetaResponse' }
        default: { $ref: '#/components/responses/SigningBetaError' }
...
components:
...
  securitySchemes:
    notificationPreferenceToken:
      type: http
      scheme: bearer
      description: Signed recipient identity from an email preference link
    notarySession:
      type: http
      scheme: bearer
      description: Notary SESSION token from POST /notary/session/verify.
    clientToken:
      type: http
      scheme: bearer
      description: Per-bid CLIENT token returned once by POST /bids (`sub` = bid id).
    supportToken:
      type: http
      scheme: bearer
      description: >-
        Per-thread live-chat token (ADR 0026): SUPPORT scope for the widget
        (returned by POST /support/messages), SUPPORT_OP scope for the
        operator's emailed reply link. `sub` = thread id in both.
  responses:
    SigningBetaError:
      description: >-
        400 malformed request; 401 absent/expired token or recent mailbox verification required;
        403 retained ownership or active notary required; 404 unknown session;
        409 stale revision, wrong stage, expired session or missing media presence;
        422 invalid cryptographic proof; 429 rate limit; 503 disabled/unavailable.
        Rate-limit storage failures fail closed. No tokens, keys or provider internals are returned.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    SigningBetaCapabilities:
      type: object
      required: [enabled, mode, legalSignatureAvailable, identityProofingAvailable, recordingAvailable, turnConfigured]
      properties:
        enabled: { type: boolean }
        mode: { type: string, const: rehearsal }
        legalSignatureAvailable: { type: boolean, const: false }
        identityProofingAvailable: { type: boolean, const: false }
        recordingAvailable: { type: boolean, const: false }
        authentication: { type: string, const: recent_email_verification }
        mediaEncryption: { type: string, const: webrtc_dtls_srtp }
        protectsAgainstMaliciousWebPublisher: { type: boolean, const: false }
        turnConfigured: { type: boolean }
        sessionDurationMs: { type: integer }
        presenceFreshnessMs: { type: integer }
    SigningBetaCommand:
      type: object
      required: [bidId, dateISO]
      properties:
        bidId: { type: string, pattern: '^[a-zA-Z0-9_-]{1,100}$' }
        dateISO: { type: string, format: date }
        revision: { type: integer, minimum: 1 }
        action: { type: string, enum: [join, admit, review, release, pause, acknowledge, close] }
        publicKeyJwk:
          type: object
          required: [kty, crv, x, y]
          properties:
            kty: { type: string, const: EC }
            crv: { type: string, const: P-256 }
            x: { type: string }
            y: { type: string }
        observationConfirmed: { type: boolean }
        connected: { type: boolean }
        type: { type: string, enum: [offer, answer] }
        sdp: { type: string, maxLength: 24000 }
        signature: { type: string, pattern: '^[a-zA-Z0-9_-]{86}$' }
    SigningBetaResponse:
      type: object
      required: [role, session, capabilities, iceServers]
      properties:
        role: { type: string, enum: [notary, client] }
        session:
          type: [object, 'null']
          properties:
            id: { type: string }
            revision: { type: integer }
            status: { type: string, enum: [waiting, admitted, reviewed, released, paused, complete, closed, expired] }
            expiresAt: { type: integer }
            document: { type: object }
            participants: { type: object }
            peerSignal: { type: [object, 'null'] }
            challenge: { type: [object, 'null'] }
        capabilities: { $ref: '#/components/schemas/SigningBetaCapabilities' }
        iceServers: { type: array, items: { type: object } }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
          enum:
            - service_inconnu
            - montant_invalide
            - date_invalide
            - date_passee
            - sous_prix_depart
            - plafond_depasse
            - mois_invalide
            - json_invalide
            - courriel_invalide
            - signature_invalide
            - introuvable
            - non_autorise
            - interdit
            - deja_retenue
            - offre_non_ouverte
            - proposition_inferieure
            - message_trop_long
            - documents_requis
            - document_inconnu
            - proposition_introuvable
            - proposition_close
            - dossier_invalide
            - parametre_requis
            - telephone_invalide
            - nom_invalide
            - etude_invalide
            - adresse_invalide
            - alertes_invalides
            - alerte_rythme_invalide
            - profil_incomplet
            - trop_de_messages
            - type_inconnu
            - code_invalide
            - code_deja_pris
            - lien_invalide
            - compte_requis
            - message_requis
            - offre_annulee
            - corps_trop_grand
            - id_manquant
            - acte_non_autorise
            - offre_non_retenue
            - expediteur_invalide
        message:
          type: string
          description: Human-readable message (fr-CA).
        manquants:
          type: array
          description: >-
            Only on `profil_incomplet` (ADR 0033): the contact fields the
            notary still has to fill before retaining, with their fr-CA labels.
          items:
            type: object
            required: [id, label]
            properties:
              id:
                type: string
                enum: [nom, telephone, adresse]
              label:
                type: string
...
    ErrorResponse:
      type: object
      required: [errors]
      properties:
        errors:
          type: array
          items:
            $ref: "#/components/schemas/Error"